feat: revert mcp
