+++ conf/app.conf
@@ -4,7 +4,7 @@ runmode = dev
 SessionOn = true
 copyrequestbody = true
 driverName = mysql
-dataSourceName = root:123456@tcp(localhost:3306)/
+dataSourceName = root:123456@tcp(localhost:23457)/
 dbName = casibase
 redisEndpoint =
 guacamoleEndpoint = 127.0.0.1:4822
@@ -12,12 +12,12 @@ isDemoMode = false
 disablePreviewMode = false
 logPostOnly = true
 landingFolder =
-casdoorEndpoint = https://door.casdoor.com
+casdoorEndpoint = http://localhost:7001
 ; casdoorEndpoint = http://localhost:8000
-clientId = af6b5aa958822fb9dc33
-clientSecret = 8bc3010c1c951c8d876b1f311a901ff8deeb93bc
-casdoorOrganization = "casbin"
-casdoorApplication = "app-casibase"
+clientId = 62706eeed086463c107c
+clientSecret = af37455c0bdec6896224a35a6a8e8cce31cb90b3
+casdoorOrganization = "built-in"
+casdoorApplication = "app-built-in"
 redirectPath = /callback
 cacheDir = "C:/casibase_cache"
 appDir = ""
